+++ .gitignore
@@ -1,3 +1,3 @@
-
+node_modules
 Frontend/app/package-lock.json
 Frontend/app/package.json
