+++ Backend/texts.csv
@@ -58,3 +58,18 @@ What should I eat for dinner today?
 Insulin production began after insulin was first discovered in 1921.
 "In 1986, Eli Lilly and Gennentech produced Humulin, the first recombinant insulins to be approved for general medical use, by using a new technology called a proinsulin route. Proinsulin route produces a single chemically synthesized sequence for native human proinsulin. This method was preferred over a chain combination procedure because it was more efficient as it required only one fermentation and subsequent purification protocol."
 I play soccer.
+i just moved to california
+['last night was fun']
+last night was fun
+last night was fun
+my name is nawid
+my name is nawid
+georgia tech is the best school
+georgia tech is the best school
+georgia tech is the best school
+georgia tech is the best school
+georgia tech is the best school
+georgia tech is the best school
+georgia tech is the best school
+georgia tech is the best school
+georgia tech is the best school
